Conflicts:
	.mailmap
	arch/arm/common/pl330.c
	arch/arm/include/asm/hardware/pl330.h
	drivers/gpio/pca953x.c
	drivers/media/video/sr030pc30.c
	drivers/net/wireless/125_109_bcmdhd/include/proto/bcmevent.h
	drivers/net/wireless/bcmdhd/bcmwifi.c
	drivers/net/wireless/bcmdhd/include/bcmspibrcm.h
	drivers/net/wireless/bcmdhd/include/bcmwifi.h
	drivers/net/wireless/bcmdhd/include/htsf.h
	drivers/net/wireless/bcmdhd/include/spid.h
	drivers/net/wireless/bcmdhd_4334/src/dhd/sys/dhd_common.c
	drivers/net/wireless/bcmdhd_4334/src/include/bcmendian.h
	drivers/net/wireless/bcmdhd_4334/src/include/bcmutils.h
	drivers/net/wireless/bcmdhd_4334/src/include/sbconfig.h
	drivers/net/wireless/bcmdhd_4334/src/include/sbhnddma.h
	drivers/net/wireless/bcmdhd_4334/src/shared/hndpmu.c
	drivers/net/wireless/bcmdhd_43362/Kconfig.old
	drivers/net/wireless/bcmdhd_43362/include/bcmpcispi.h
	drivers/net/wireless/bcmdhd_43362/include/bcmsdstd.h
	drivers/net/wireless/bcmdhd_43362/include/sbsdio.h
	drivers/net/wireless/bcmdhd_43362/wl_iw.c
	drivers/net/wireless/bcmdhd_lnx/aiutils.c
	drivers/net/wireless/bcmdhd_lnx/bcmsdh.c
	drivers/net/wireless/bcmdhd_lnx/bcmsdh_linux.c
	drivers/net/wireless/bcmdhd_lnx/bcmsdh_sdmmc.c
	drivers/net/wireless/bcmdhd_lnx/bcmsdh_sdmmc_linux.c
	drivers/net/wireless/bcmdhd_lnx/bcmutils.c
	drivers/net/wireless/bcmdhd_lnx/bcmwifi.c
	drivers/net/wireless/bcmdhd_lnx/dhd_bus.h
	drivers/net/wireless/bcmdhd_lnx/dhd_custom_gpio.c
	drivers/net/wireless/bcmdhd_lnx/dhd_dbg.h
	drivers/net/wireless/bcmdhd_lnx/dhd_linux_sched.c
	drivers/net/wireless/bcmdhd_lnx/dhd_proto.h
	drivers/net/wireless/bcmdhd_lnx/dhd_sdio.c
	drivers/net/wireless/bcmdhd_lnx/include/aidmp.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmcdc.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmperf.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmsdbus.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmsdh.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmsdh_sdmmc.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmsdpcm.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmsdspi.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmspi.h
	drivers/net/wireless/bcmdhd_lnx/include/bcmwifi.h
	drivers/net/wireless/bcmdhd_lnx/include/dhdioctl.h
	drivers/net/wireless/bcmdhd_lnx/include/hndrte_armtrap.h
	drivers/net/wireless/bcmdhd_lnx/include/hndrte_cons.h
	drivers/net/wireless/bcmdhd_lnx/include/hndsoc.h
	drivers/net/wireless/bcmdhd_lnx/include/htsf.h
	drivers/net/wireless/bcmdhd_lnx/include/linux_osl.h
	drivers/net/wireless/bcmdhd_lnx/include/linuxver.h
	drivers/net/wireless/bcmdhd_lnx/include/miniopt.h
	drivers/net/wireless/bcmdhd_lnx/include/msgtrace.h
	drivers/net/wireless/bcmdhd_lnx/include/osl.h
	drivers/net/wireless/bcmdhd_lnx/include/packed_section_end.h
	drivers/net/wireless/bcmdhd_lnx/include/packed_section_start.h
	drivers/net/wireless/bcmdhd_lnx/include/pcicfg.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/802.11_bta.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/802.11e.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/802.1d.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/bcmeth.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/bcmip.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/eapol.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/ethernet.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/sdspi.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/vlan.h
	drivers/net/wireless/bcmdhd_lnx/include/proto/wpa.h
	drivers/net/wireless/bcmdhd_lnx/include/sbpcmcia.h
	drivers/net/wireless/bcmdhd_lnx/include/sbsdpcmdev.h
	drivers/net/wireless/bcmdhd_lnx/include/sbsocram.h
	drivers/net/wireless/bcmdhd_lnx/include/sdio.h
	drivers/net/wireless/bcmdhd_lnx/include/sdioh.h
	drivers/net/wireless/bcmdhd_lnx/include/sdiovar.h
	drivers/net/wireless/bcmdhd_lnx/include/siutils.h
	drivers/net/wireless/bcmdhd_lnx/include/trxhdr.h
	drivers/net/wireless/bcmdhd_lnx/include/typedefs.h
	drivers/net/wireless/bcmdhd_lnx/sbutils.c
	drivers/net/wireless/bcmdhd_lnx/siutils.c
	drivers/net/wireless/bcmdhd_lnx/siutils_priv.h
	drivers/net/wireless/bcmdhd_lnx/wl_dbg.h
	drivers/net/wireless/bcmdhd_lnx/wl_iw.h
	drivers/tty/serial/8250.c
	drivers/usb/otg/otg_id.c
	drivers/video/display/Kconfig
	drivers/video/display/Makefile
	fs/partitions/check.c
	kernel/sched.c

+++ drivers/usb/core/message.c
@@ -308,7 +308,8 @@ static void sg_complete(struct urb *urb)
 				retval = usb_unlink_urb(io->urbs [i]);
 				if (retval != -EINPROGRESS &&
 				    retval != -ENODEV &&
-				    retval != -EBUSY)
+				    retval != -EBUSY &&
+				    retval != -EIDRM)
 					dev_err(&io->dev->dev,
 						"%s, unlink --> %d\n",
 						__func__, retval);
@@ -317,7 +318,6 @@ static void sg_complete(struct urb *urb)
 		}
 		spin_lock(&io->lock);
 	}
-	urb->dev = NULL;
 
 	/* on the last completion, signal usb_sg_wait() */
 	io->bytes += urb->actual_length;
@@ -435,7 +435,7 @@ int usb_sg_init(struct usb_sg_request *io, struct usb_device *dev,
 
 			len = sg->length;
 			if (length) {
-				len = min_t(unsigned, len, length);
+				len = min_t(size_t, len, length);
 				length -= len;
 				if (length == 0)
 					io->entries = i + 1;
@@ -524,7 +524,6 @@ void usb_sg_wait(struct usb_sg_request *io)
 		case -ENXIO:	/* hc didn't queue this one */
 		case -EAGAIN:
 		case -ENOMEM:
-			io->urbs[i]->dev = NULL;
 			retval = 0;
 			yield();
 			break;
@@ -542,7 +541,6 @@ void usb_sg_wait(struct usb_sg_request *io)
 
 			/* fail any uncompleted urbs */
 		default:
-			io->urbs[i]->dev = NULL;
 			io->urbs[i]->status = retval;
 			dev_dbg(&io->dev->dev, "%s, submit --> %d\n",
 				__func__, retval);
@@ -593,7 +591,10 @@ void usb_sg_cancel(struct usb_sg_request *io)
 			if (!io->urbs [i]->dev)
 				continue;
 			retval = usb_unlink_urb(io->urbs [i]);
-			if (retval != -EINPROGRESS && retval != -EBUSY)
+			if (retval != -EINPROGRESS
+					&& retval != -ENODEV
+					&& retval != -EBUSY
+					&& retval != -EIDRM)
 				dev_warn(&io->dev->dev, "%s, unlink --> %d\n",
 					__func__, retval);
 		}
@@ -1135,8 +1136,6 @@ void usb_disable_interface(struct usb_device *dev, struct usb_interface *intf,
  * Deallocates hcd/hardware state for the endpoints (nuking all or most
  * pending urbs) and usbcore state for the interfaces, so that usbcore
  * must usb_set_configuration() before any interfaces could be used.
- *
- * Must be called with hcd->bandwidth_mutex held.
  */
 void usb_disable_device(struct usb_device *dev, int skip_ep0)
 {
@@ -1189,7 +1188,9 @@ void usb_disable_device(struct usb_device *dev, int skip_ep0)
 			usb_disable_endpoint(dev, i + USB_DIR_IN, false);
 		}
 		/* Remove endpoints from the host controller internal state */
+		mutex_lock(hcd->bandwidth_mutex);
 		usb_hcd_alloc_bandwidth(dev, NULL, NULL, NULL);
+		mutex_unlock(hcd->bandwidth_mutex);
 		/* Second pass: remove endpoint pointers */
 	}
 	for (i = skip_ep0; i < 16; ++i) {
@@ -1749,7 +1750,6 @@ free_interfaces:
 	/* if it's already configured, clear out old state first.
 	 * getting rid of old interfaces means unbinding their drivers.
 	 */
-	mutex_lock(hcd->bandwidth_mutex);
 	if (dev->state != USB_STATE_ADDRESS)
 		usb_disable_device(dev, 1);	/* Skip ep0 */
 
@@ -1762,6 +1762,7 @@ free_interfaces:
 	 * host controller will not allow submissions to dropped endpoints.  If
 	 * this call fails, the device state is unchanged.
 	 */
+	mutex_lock(hcd->bandwidth_mutex);
 	ret = usb_hcd_alloc_bandwidth(dev, cp, NULL, NULL);
 	if (ret < 0) {
 		mutex_unlock(hcd->bandwidth_mutex);
@@ -1836,106 +1837,6 @@ free_interfaces:
 	if (cp->string == NULL &&
 			!(dev->quirks & USB_QUIRK_CONFIG_INTF_STRINGS))
 		cp->string = usb_cache_string(dev, cp->desc.iConfiguration);
-/* Uncomment this define to enable the HS Electrical Test support */
-#define DWC_HS_ELECT_TST 1
-#ifdef DWC_HS_ELECT_TST
-		/* Here we implement the HS Electrical Test support. The
-		 * tester uses a vendor ID of 0x1A0A to indicate we should
-		 * run a special test sequence. The product ID tells us
-		 * which sequence to run. We invoke the test sequence by
-		 * sending a non-standard SetFeature command to our root
-		 * hub port. Our dwc_otg_hcd_hub_control() routine will
-		 * recognize the command and perform the desired test
-		 * sequence.
-		 */
-		if (dev->descriptor.idVendor == 0x1A0A) {
-			/* HSOTG Electrical Test */
-			dev_warn(&dev->dev, "VID from HSOTG Electrical Test Fixture\n");
-
-			if (dev->bus && dev->bus->root_hub) {
-				struct usb_device *hdev = dev->bus->root_hub;
-				dev_warn(&dev->dev, "Got PID 0x%x\n", dev->descriptor.idProduct);
-
-				switch (dev->descriptor.idProduct) {
-				case 0x0101:	/* TEST_SE0_NAK */
-					dev_warn(&dev->dev, "TEST_SE0_NAK\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x300, NULL, 0, HZ);
-					break;
-
-				case 0x0102:	/* TEST_J */
-					dev_warn(&dev->dev, "TEST_J\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x100, NULL, 0, HZ);
-					break;
-
-				case 0x0103:	/* TEST_K */
-					dev_warn(&dev->dev, "TEST_K\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x200, NULL, 0, HZ);
-					break;
-
-				case 0x0104:	/* TEST_PACKET */
-					dev_warn(&dev->dev, "TEST_PACKET\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x400, NULL, 0, HZ);
-					break;
-
-				case 0x0105:	/* TEST_FORCE_ENABLE */
-					dev_warn(&dev->dev, "TEST_FORCE_ENABLE\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x500, NULL, 0, HZ);
-					break;
-
-				case 0x0106:	/* HS_HOST_PORT_SUSPEND_RESUME */
-					dev_warn(&dev->dev, "HS_HOST_PORT_SUSPEND_RESUME\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x600, NULL, 0, 40 * HZ);
-					break;
-
-				case 0x0107:	/* SINGLE_STEP_GET_DEVICE_DESCRIPTOR setup */
-					dev_warn(&dev->dev, "SINGLE_STEP_GET_DEVICE_DESCRIPTOR setup\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x700, NULL, 0, 40 * HZ);
-					break;
-
-				case 0x0108:	/* SINGLE_STEP_GET_DEVICE_DESCRIPTOR execute */
-					dev_warn(&dev->dev, "SINGLE_STEP_GET_DEVICE_DESCRIPTOR execute\n");
-					usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT,
-							USB_PORT_FEAT_TEST, 0x800, NULL, 0, 40 * HZ);
-					break;
-
-#ifdef CONFIG_USB_OTG
-				case 0x0200: /* TEST DEVICE REQUIRED BY COMPLIANCE TEST */
-					dev_warn(&dev->dev, "TEST DEVICE REQUIRED BY COMPLIANCE TEST\n");
-					hcd->self.otg_vbus_off = dev->descriptor.bcdDevice & 0x01;
-					if (hcd->self.otg_vbus_off)
-						usb_control_msg(hdev, usb_sndctrlpipe(hdev, 0),
-							USB_REQ_SET_FEATURE, USB_RT_PORT, USB_PORT_FEAT_TEST, 0x1000,
-							NULL, 0, 1000);
-					else if (hcd->self.is_b_host) {
-						/* Suspend within TTST_SUSP after HNP */
-						usb_host_suspend_test_device(dev);
-					} else
-						schedule_delayed_work(&dev->bus->maint_conf_session_for_td,
-							msecs_to_jiffies(HOST_VBOFF));
-					break;
-#endif
-				default:
-					dev_warn(&dev->dev, "UNKNOWN PID from test fixture\n");
-					break;
-				}
-			}
-		}
-#endif /* DWC_HS_ELECT_TST */
 
 	/* Now that all the interfaces are set up, register them
 	 * to trigger binding of drivers to interfaces.  probe()
